fixed fence planning - dfs
i tried to debug my code after getting this question wrong. I kept reading it time and time again and couldnt figure out why it wasn't working. I rewrote the code, and the same thing happened. This code was much cleaner, so I could easily debug. The mistake was such a simple one, as shown in the edits. I copied the old code and fixed it. I am leaving the fixed long code, since that was my actual first and basically correct attempt.