Birthday Party - dfs on adj matrix
I originally attempted this problem on a adj list, but realized I had to do extra work to remove edges. With adj matrix's, I just mark them as false. So, if given the edge of nodes, (a,b), I just mark the edge between a,b in the adj matrix as false and dfs, checking if there is a single connected component.