p2 - nesting depth
notice that the number of number of '(' is equal to a[i] - a[i-1], if neg then zero. Also, notice that number of ')' is equal to a[i] - a[i+1], if neg than zero. Then, the rest is trivial. Another method, is it just add the necessary parenthesis, then go through and remove all ")(" in the ans string. I tried this at first, but ended up deleting that code, as I thought I could write this one faster.