The Great Revegetation - bi-dfs
For this problem, using some math and knowledge of components, you will realize, 2^k, where k is connected components, is the answer. If you found this, the fact that the answer is required in binary, is a giant hint lol.  Turns out I got a test case wrong. I am almost 100% certain it is because the graph given is impossible to bicolor. I will hopefully fix this code tomorrow.