icy perimeter - floodfill
Find number of connected components using flood fill. Then count number of hastags in that component, and find the perimeter of it. Honestly, if you know flood fill, the only hard portion of this problem is finding the perimeter. It was a little annoying having to check all 4 surrounding boxes. There may be better ways that I did not realize.