cf #577 div 2 C - bin search on ans
First, realize that the number of moves needed to get to a certain median is (k = n/2) sigma (n) of max(0, x- arr[k]). After, bin search on ans by testing data sets between 1 and 2e9. I originally made a mistake by doing least + (greatest - least)/2, because if left == 0 and right == 1, this results in a infinite loop. The fix was to add one, to have the mid == 1, which would cause the code to terminate, giving us our answer.