Cow Dance Show - bsearch on ans
This problem isn't very hard in the binary searching on the answer category. This problem is obviously seen as a bsearch on ans kind of problem. I had a harder time trying to find out a way to see if a value of k works. I was constantly trying to subtract other values to count the current moment in time, when I could just add to other elements, so they could reflect themselves, respective to the amount of time has passed. A priority queue makes this problem possible. You just need to use a function pointer, or one of the built in std ones like greater<int>.