Paired Up - two pointer
save the values into some storage data structure. Sort them by units of time to milk. After, look for unique times only. So, minus the min(number of times1, number of times2), to account for the other min-1 times not being unique and being a waste of time. This makes the code fast enough.