Hay Bayle - prefix sums
Something important about prefix sums is that you should notice that outside the sub array sum, you have a lot of power to change and manipulate the data. In this problem, the main idea was finding out a faster (than O(N^2)) way to assign values to a sub array. Something to notice and know is that, say you want to increase a_x -> a_n +1 per element. If you increase a_x  by 1, and decrease a_(n+1) by 1, the prefix a_x to a_n will all be increased by 1. Do this for all the instructions, make a prefix sum, sort it, and print out arr[n/2] for the median !