Max Subarray Sum - prefix sums
Notice that the max prefix sum - the min prefix sum is normally the answer. There are just a few things to take notice of. You have to use 0, when you start, checking the entire sub array, but after you want to use negatives if possible for a greater maxpf. Otherwise, just search for maxpf, and the answer is the greatest maxpf found.