High Card Wins - Greedy
This problem is greedy. Notice that it is best to always choose the first card greater than the opponents. This can be proven with examples, one example is some like 1,2, 5 - opponent, 3, 4, 6 you. You start with their smallest card. Look for the smallest card that can beat theirs, this way you save the better cards for an unforeseen future. Next,  do the same thing on their next card.