High Card Low Card - Greedy
I recently did the silver variation, and I can say they are almost no different. They are problems from the very first contest that included platinum, which makes them much easier. For this problem, the greedy algo is the same as the silver variation. The only difference is we want to look at the vectors split in to two. We simply just need to know how pointers work and this problem is just as easy as the silver version.