Digit Code Forces Problem - Greedy
Cool greedy problem. It is easy to find max, just keep using greatest possible number. I think I used a terrible way of finding the minimum, but its the first thing I thought of. I reversed the numbers, and check if any zeros. If there is, I find the first non zero, and subtract it to account for first digit being 1. This is bad, because it fails on 1 0, so I had to bluntly check this in the beginning, which is not fluid and ugly. My digit array is also a waste of stack space, there are other ways to get char version of int with just ascii values. besides that, its pretty intuitive.