Convention - Bin search on answer
Binary search on answer problems are pretty fun. It is like brute forcing, but with a tiny advantage. For this problem, we literally just brute force. We plug in a time, and simulate it. The only difference is the fact that instead of going from something like 1 -> N, we binary search a subset of like 0 -> 1 billion or something (it really doesn't matter too much cuz its logn, it should just be big). Besides that, I believe my use of upper_bound was nice. It saved me some time, but I had some trouble with checking if my address was garbage memory.