Two Button - DFS or math
there is a graph solution (or tree to be more specific, bin tree even) where each node leads to the path of dividing by 2 or adding 1. I only noticed after reading the solution. My solution was based on observation. I at first tried to go from a to b. I later realized it was much simpler to multiply by 2 and subtract 1. So I went from B to A, and sure enough it was much simpler. Divide if even, else add. count operations. I am more interested in the graph problem and am going to look for another user who tried that approach.