Phoenix And Beauty -  Observation?
These questions are some of the ones I hate the most. They are like borderline greedy, but just random. They are not entirely ad hoc, but it pretty much is. For this problem, you just need to notice that repeating any subarray of the distinct elements will make the exact same sums. They constraints were luckily small, so it made it a little easier. Also, just notice that it is impossible to construct the beautiful array if the number of distinct elements is greater than the sub array length for apparent reasons (this is different if 0 is included, but its not.)