Vacations - Greedy
We keep track of what is behind us. If the number behind us is equal to us and isnt 3, or the current number is 0, we add one to our answer. If we got a 3, we just need to make it 2 if 1, and 1 if 2. Otherwise, it will eventually become a 0. Update behind other wise.