Flight Routes Check - dfs or bfs
Cool problem. I actually wasnt sure how to begin and looked for some clues online. One said to know this theorem, "The theorem used here is that if one vertex can both reach and be reached by all others, then every vertex in this graph can reach all others." This means that if one vertex can be reached by all other vertexes, and this vertex can reach all other vertexes, all connections are possible. From here on, its easier. We just need to reverse the graph and check if this is true. If it is, we print yes, otherwise, some node was not reached in the reversed graph, or the original graph. Check the visited array. The node that isnt visited is false, so just print that one along with what ever node u dfs(ed) from (1 is best choice). *NOTE* Remember in the reversed graph, 1->x is really x->1, so make sure to print the node that isnt reached from 1 reversed.