Number Game - Case Work
Definitely not a fun problem. I had to spend tons of time looking at the prime factors, but still couldn't figure it out. I read the beginning of the solution. The only cases in which FastestFinger can win is if n == 1, n is in the form 2^x, with x > 1, or n being in the form 2 * prime number(not 2). I threw in the odd case so I wouldnt have to check evenness later.