News Distribution - dfs
This is a common graph problem, kinda. The goal of this problem is (construct a graph *) to identify the number of nodes in every connected component. This needs to be done during the dfs or you will time limit exceed. The way I did it was by coloring the visited nodes. Instead of making them just true or false, I had 0 represent not visited, and any other number represent their group number. To get the answers for every node, if this node is not visited (first iteration and others), I dfs from that node, and all those nodes get marked as group 1. Later, if any groups are visited and are in group 1, I print the answer for that we got from that original node. If it isnt in group 1, I repeat.