Diamond Collector - 2p, but i used bin search
It turns out separating the buckets into one bucket with most possible, and the other with next most does not work. I can see why, but I cant understand any of the other solutions at the moment. Ill push this for now until it is changed. This code fails on cases 5,6, and 10.