Greedy
The way i viewed this problem was that I was going to have a min of (#movies, k), at first. So, if possible, I want to reuse the people who have already seen movies before using my special (watch any movie) people. Sorting by end is the first greedy part (check movie fest 1 for that greedy). For this problem, we use a multiset with binary searching. For every start time, if there is an end time that is less than it, we want to use that, other wise, it is optimal to use our special people (sorted by end time). Upper bound should be used here because of the fact that it returns an iterator to the first element greater (which means the previous must be <=), but if we get a pointer to the first element, it would mean no element was greater and we want to skip this value because we cant do anything with it.