Math
Kinda weird kinda cool. I kinda guessed this based on the fact that 180 degrees was 2*r as given in the example. It can easily be proved that is optimal to use 2*r translations. Basically just realize that the distance between 2 points / (2*r) is our answer. We should ceil this in case it needs an extra move with a left over (we would not fully use the rotation).