Prefix Sums
After changing to prefix sums version of array, notice that to achieve a good subarray a_i - i = a_j - j, easily find this with some math. So, just count the number of teams a_i - i is found. this number choose 2 is the number of subarrays for that possible score.