Cycles ? Greedy ? Math ?
After finding how much is needed to make the number divisible by k, we realize that we will iterate the maximum number of times a number appears and go directly to that number. All we need to do is determine which number appears the most number of times and how many times. I was trying something else at first and used an unordered_map and ended up keeping it for this method and it surprisingly time limited exceeded (trashy hash key bruhhhh), so i changed to map and it worked totally fine. A dynamic array would work fine here as long as it was 1e9 size i think.