binary search on answer
pretty self explanatory. I was attempting to approach this like a rates math problem and just calculate the amount robots made per second. This would give us a O(n) answer would be cool. Only problem is that the time needed is pretty large and i was getting overflows with doubles and precision stuff so i just did the normal bsearch.