attempting to fix code
fails on second half of test cases. I dont know why. I assume its because of the map, its a very inefficient way of checking that I use for the graph. The other option was like a 3d vector, but i thought it would be big brain to try this. I may attempt to fix this again.