Splitting The Field - sweep line
This is one of those rare problems where finding the solution on paper is easier than writing the code. Of course, I would rather deal with implementation any day, but this one was p hard. The idea behind the problem is to sort the values by x first and create a vertical line that sweeps through every single cow point. You would only travel on cow points because u would want to tighten the fence. When you move this vertical line, the cow you are on becomes a part of the left fence, and the rest are part the right fence. After, you reflect this by sorting by y coordinate making it seem as if you are sweeping horizontally. Multisets and iterators are used frequently. I had to look at the compression of the id and coordinates as well, since i was trying something diff.