i have nothing to say about this problem
This problem took SO damn long. It is really hard so I didnt really expect to come up with the solution on my own, but what the hell. Even just understanding the solution is hard. The math at the beginning is super annoying. Its basically an mid level amc question with grunt work mixed in with annoying implementation. After rereading a thousand diff solutions, I somewhat understood what to do.