Math and sorting
This is the furthest I have been able to solve a (usaco.guide rated) hard problem. I almost had it on my own, but i was confused about the left and right portion. I knew to move to the left, its only possible 1 at a time, but I wasnt sure if numbers to the right could also be a max, since they move to their position in one iteration.  I read that they should be shown as negative and positive and that made sense to me.