observation and sorting
honestly surprised it worked. My initial thoughts were to just append the latest character to make all the strings have the same size, but i couldnt find a way to do this in time (but it should work). Then, I realized I could basically brute force this. When choosing which strings to use, would I want to add string a first, or string b first. Well, just add them in that order and use the smaller (lexicographically) one.  This ended up working. The solution has an actual proof, but i didnt really understand it, but i also didnt thoroughly study it. This is the highest rated problem ive done so far (1700), but it def didnt feel harder 1500.