redid, solved correctly
we store the maximum diamonds we can hold, given that a_i is chosen. Then, we find the max diamonds possible, if we do not chose our current amount. This works because it guarantees not intersections. If we didnt care about intersections, the answer is just max of diamonds held if given we choose a_i to be in (top 2), but what we do is choose one group, and find the largest group that is not that one.