just a test of pointers and how well u can remember what to do
As we go through all the times in sorted order, we store the current max, and the number of cows with that max. We then simulate moving the cow to another gallon group. We change if the past winner if diff from now (both ways) or if we stay winning, but have other people with us (found by checking if prev count or new count is 1).