mostly just pointer usage from set
start with the largest boot, when the depths are sorted from largest to smallest (for boots and snow), and remove any snow depths that are greater than our depth, and see the index before that and after in our index list to see how many boots we skipped. As we go down the list we will only change a max of n, while going through b times.