used queue, prob math sol
main idea: if a end > b end, (a start < b start), then a might as well end at the b end point, and b at a end. The reason for this is because the distance is squared, and as we know, if we graph x^2, its rate of increase at each point is 2x, by the d/dx. (idk if that was even what i was saying, i just wanted to say the difference between squares is increasing because the velocity of the squares is not constant.) i wonder if theres any way to know the optimal starting position (maybe where ever the most cows are?).